+++ packages/core/avatar/src/components/AvatarImage.js
@@ -2,6 +2,7 @@
 
 import GlobalTheme from '@atlaskit/theme';
 import React, { PureComponent } from 'react';
+import { canUseDOM } from 'exenv';
 import { Slot, ShapeGroup, Svg } from '../styled/AvatarImage';
 import type { AppearanceType, SizeType } from '../types';
 
@@ -78,7 +79,7 @@ export const clearCache = () => {
 export default class AvatarImage extends PureComponent<Props, State> {
   state: State = {
     hasError: false,
-    // if provided a scr - we need to load it
+    // if provided a src - we need to load it
     isLoading: Boolean(this.props.src),
   };
   isComponentMounted: boolean;
@@ -87,12 +88,14 @@ export default class AvatarImage extends PureComponent<Props, State> {
     this.isComponentMounted = true;
     this.loadImage();
   }
+
   // handle case where `src` is modified after mount
   componentWillReceiveProps(nextProps: Props) {
     if (nextProps.src && this.props.src !== nextProps.src) {
       this.setState({ isLoading: true });
     }
   }
+
   componentDidUpdate(prevProps: Props) {
     if (this.props.src && this.props.src !== prevProps.src) {
       this.loadImage();
@@ -101,6 +104,7 @@ export default class AvatarImage extends PureComponent<Props, State> {
   componentWillUnmount() {
     this.isComponentMounted = false;
   }
+
   loadImage = () => {
     // nothing to load
     if (!this.props.src) {
@@ -112,24 +116,30 @@ export default class AvatarImage extends PureComponent<Props, State> {
     img.onerror = this.handleLoadError;
     img.src = this.props.src;
   };
+
   handleLoad = (hasError: boolean) => {
     if (this.isComponentMounted) {
       this.setState({ hasError, isLoading: false });
     }
   };
+
   handleLoadSuccess = () => {
     if (typeof this.props.src === 'string') {
       cache[this.props.src] = true;
     }
     this.handleLoad(false);
   };
-  handleLoadError = () => this.handleLoad(true);
+
+  handleLoadError = () => {
+    this.handleLoad(true);
+  };
 
   render() {
     const { alt, src, appearance, size } = this.props;
     const { hasError, isLoading } = this.state;
     const showDefault = !isLoading && (!src || hasError);
-    const imageUrl: ?string = src && (!isLoading || cache[src]) ? src : null;
+    const imageUrl: ?string =
+      src && (!isLoading || cache[src] || !canUseDOM) ? src : null;
     return showDefault ? (
       <DefaultImage
         appearance={appearance}