+++ packages/editor/editor-core/src/__tests__/unit/plugins/gap-cursor/_utils.ts
@@ -54,7 +54,6 @@ export const leafBlockNodes = {
   mediaSingle: mediaSingle({ layout: 'center' })(
     media({
       id: temporaryFileId,
-      __key: temporaryFileId,
       type: 'file',
       collection: testCollectionName,
       width: 100,