+++ packages/editor/editor-core/src/__tests__/visual-regression/table/block-alignment.ts
@@ -18,24 +18,25 @@ describe('Table with block looks correct for fullpage:', () => {
     page = global.page;
   });
 
+  afterEach(async () => {
+    await snapshot(page, 0.01);
+  });
+
   it('default layout ', async () => {
     await initFullPageEditorWithAdf(page, adf, Device.LaptopMDPI);
     await page.click(getSelectorForTableCell({ row: 4, cell: 1 }));
-    await snapshot(page);
   });
 
   it('wide layout ', async () => {
     await initFullPageEditorWithAdf(page, adf, Device.LaptopMDPI);
     await page.click(getSelectorForTableCell({ row: 4, cell: 1 }));
     await setTableLayout(page, 'wide');
-    await snapshot(page);
   });
 
   it('full-width layout ', async () => {
     await initFullPageEditorWithAdf(page, adf, Device.LaptopHiDPI);
     await page.click(getSelectorForTableCell({ row: 4, cell: 1 }));
     await setTableLayout(page, 'full-width');
-    await snapshot(page);
   });
 });
 
@@ -47,9 +48,12 @@ describe('Table with block looks correct for comment:', () => {
     page = global.page;
   });
 
+  afterEach(async () => {
+    await snapshot(page, 0.01);
+  });
+
   it('default layout ', async () => {
     await initCommentEditorWithAdf(page, adf, Device.LaptopMDPI);
     await page.click(getSelectorForTableCell({ row: 4, cell: 1 }));
-    await snapshot(page);
   });
 });