+++ packages/editor/editor-core/src/__tests__/visual-regression/table/numbered-table.ts
@@ -15,21 +15,22 @@ describe('Snapshot Test: numbered table', () => {
     page = global.page;
   });
 
+  afterEach(async () => {
+    await snapshot(page);
+  });
+
   it(`looks correct at LaptopMDPI for fullpage`, async () => {
     await initFullPageEditorWithAdf(page, adf, Device.LaptopMDPI);
     await page.click(getSelectorForTableCell({ row: 1, cell: 1 }));
-    await snapshot(page);
   });
 
   it(`looks correct at iPadPro for fullpage`, async () => {
     await initFullPageEditorWithAdf(page, adf, Device.iPadPro);
     await page.click(getSelectorForTableCell({ row: 1, cell: 1 }));
-    await snapshot(page);
   });
 
   it(`looks correct at LaptopMDPI for comment`, async () => {
     await initCommentEditorWithAdf(page, adf, Device.LaptopMDPI);
     await page.click(getSelectorForTableCell({ row: 1, cell: 1 }));
-    await snapshot(page);
   });
 });