+++ packages/media/media-card/src/root/inlinePlayer.tsx
@@ -1,6 +1,11 @@
 import * as React from 'react';
 import { Component } from 'react';
-import { Context, FileIdentifier } from '@atlaskit/media-core';
+import {
+  Context,
+  FileIdentifier,
+  FileState,
+  MediaFileArtifacts,
+} from '@atlaskit/media-core';
 import { Subscription } from 'rxjs/Subscription';
 import { CustomMediaPlayer } from '@atlaskit/media-ui';
 import { InlinePlayerWrapper } from './styled';
@@ -19,6 +24,25 @@ export interface InlinePlayerProps {
 export interface InlinePlayerState {
   fileSrc?: string;
 }
+
+export const getPreferredVideoArtifact = (
+  fileState: FileState,
+): keyof MediaFileArtifacts | undefined => {
+  if (fileState.status === 'processed' || fileState.status === 'processing') {
+    const { artifacts } = fileState;
+    if (!artifacts) {
+      return undefined;
+    }
+
+    return artifacts['video_1280.mp4']
+      ? 'video_1280.mp4'
+      : artifacts['video_640.mp4']
+      ? 'video_640.mp4'
+      : undefined;
+  }
+
+  return undefined;
+};
 
 export class InlinePlayer extends Component<
   InlinePlayerProps,
@@ -52,16 +76,17 @@ export class InlinePlayer extends Component<
             }
           }
 
-          if (state.status === 'processed') {
+          if (state.status === 'processed' || state.status === 'processing') {
+            const artifactName = getPreferredVideoArtifact(state);
             const { artifacts } = state;
+            if (!artifactName || !artifacts) {
+              return;
+            }
 
             try {
-              const preferedArtifact = artifacts['video_1280.mp4']
-                ? 'video_1280.mp4'
-                : 'video_640.mp4';
               const fileSrc = await context.file.getArtifactURL(
                 artifacts,
-                preferedArtifact,
+                artifactName,
                 collectionName,
               );