+++ packages/media/media-core/src/file/index.ts
@@ -306,6 +306,7 @@ export class FileFetcherImpl implements FileFetcher {
 
       subject.next({
         status: 'processing',
+        representations: {},
         ...stateBase,
       });
       subject.complete();