+++ packages/search/global-search/src/components/jira/NoResultsState.tsx
@@ -3,11 +3,13 @@ import { FormattedMessage } from 'react-intl';
 import styled from 'styled-components';
 import { messages } from '../../messages';
 import NoResults from '../NoResults';
-import { ResultItemGroup } from '@atlaskit/quick-search';
+import { ResultItemGroup, CancelableEvent } from '@atlaskit/quick-search';
 import JiraAdvancedSearch from './JiraAdvancedSearch';
+import { JiraEntityTypes } from '../SearchResultsUtil';
 
 export interface Props {
   query: string;
+  onAdvancedSearch?: (e: CancelableEvent, entity: JiraEntityTypes) => void;
 }
 
 const Container = styled.div`
@@ -18,7 +20,7 @@ const Container = styled.div`
 
 export default class NoResultsState extends React.Component<Props> {
   render() {
-    const { query } = this.props;
+    const { query, onAdvancedSearch } = this.props;
 
     return (
       <>
@@ -32,6 +34,7 @@ export default class NoResultsState extends React.Component<Props> {
             <JiraAdvancedSearch
               query={query}
               analyticsData={{ resultsCount: 0, wasOnNoResultsScreen: true }}
+              onClick={onAdvancedSearch}
             />
           </Container>
         </ResultItemGroup>