+++ packages/editor/embedded-document/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/embedded-document
 
+## 0.2.5
+- Updated dependencies [4aee5f3cec](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/4aee5f3cec):
+  - @atlaskit/editor-common@33.0.0
+  - @atlaskit/editor-core@102.0.0
+  - @atlaskit/renderer@41.0.0
+
 ## 0.2.4
 - Updated dependencies [4a84fc40e0](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/4a84fc40e0):
   - @atlaskit/editor-core@101.0.0